Separate classes in single-class file #14
Inputs and rigid WIP
