+++ src/lib/rigid/src/layerMask.ts
