+++ src/lib/rigid/src/vector.ts
