+++ tsconfig.json
@@ -11,8 +11,8 @@
 		"paths": {
 			"@gandolphinnn/shared": ["src/lib/shared/index.ts"],
 			"@gandolphinnn/graphics": ["src/lib/graphics/src/index.ts"],
-			"@gandolphinnn/inputs": ["src/lib/inputs/index.ts"],
-			"@gandolphinnn/rigid": ["src/lib/rigid/index.ts"],
+			"@gandolphinnn/inputs": ["src/lib/inputs/src/index.ts"],
+			"@gandolphinnn/rigid": ["src/lib/rigid/src/index.ts"],
 			"@gandolphinnn/game": ["src/lib/game/src/index.ts"],
 			"*": ["node_modules"]
 		}
